requirements.txt
google-genai>=0.4.0
groq>=0.9.0
rich>=13.7.0
pydantic>=2.5.0
python-dotenv>=1.0.0
